+++ changes.md
@@ -1,3 +1,7 @@
+### Ver 0.0.2.2
+- Fix user and group function
+- Process logs are taken individually
+
 ### Ver 0.0.2.1
 - Fix path error
 - Fix javascript active effect